+++ main-server/src/config.js
@@ -197,3 +197,5 @@ export const masterServerPort = process.env.MASTER_SERVER_PORT || 10000;
 export const workerDisconnectedTimeout = 60*1000;
 
 export const workerCallTimeout = 30*1000;
+
+export const apiToMqTimeout = 5*1000;