+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## TBD
+
+IMPROVEMENTS:
+
+- Introduce master-worker architecture with backward-compatible option to be standalone.
+
 ## 1.0.1 (January 24, 2019)
 
 BUG FIXES:
@@ -27,7 +33,6 @@ IMPROVEMENTS:
   - `/num_processing_blocks`: Number of processing blocks
   - `/processing_blocks`: Array of processing block heights in string (e.g. `51-63`, `99`)
   - `/num_processing_inbound_messages`: Number of processing inbound MQ messages
-- Introduce master-worker architecture with backward-compatible option to be standalone.
 
 ## 1.0.0 (December 7, 2018)